pyproject.toml
[project]
name = "quantum"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "geopandas>=1.1.2",
    "numpy~=2.4.1",
    "pandas~=2.3.3",
    "plotly~=6.5.1",
    "pyyaml>=6.0.3",
    "streamlit~=1.52.2",
]
